requirements.txt
streamlit
requests
google-generativeai
python-dotenv
gTTS
tqdm
Pillow